signature-verification: package renaming and updated examples
